requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
Markdown==3.7
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==7.2.1
sqlparse==0.5.3
pillow==11.1.0